fix: TSC errors with Vite
