+++ .eslintrc.json
@@ -16,6 +16,13 @@
     "no-console": "warn",
     "quotes": ["warn", "single"],
     "react/react-in-jsx-scope": "off",
+    "@typescript-eslint/no-unused-vars": [
+      "warn",
+      {
+        "varsIgnorePattern": "^_",
+        "destructuredArrayIgnorePattern": "^_"
+      }
+    ],
     "react-refresh/only-export-components": [
       "warn",
       { "allowConstantExport": true }
