+++ src/lib/web3/hooks/useUserBankBalances.ts
@@ -53,7 +53,7 @@ function useAllUserBankBalances(): UseQueryResult<Coin[]> {
   useEffect(() => {
     if (address) {
       const onTxBalanceUpdate = (
-        event: MessageActionEvent,
+        _event: MessageActionEvent,
         tx: TendermintTxData
       ) => {
         const events = tx.value.TxResult.result.events.map(mapEventAttributes);
