+++ src/pages/Orderbook/OrderbookChart.tsx
@@ -179,9 +179,9 @@ export default function OrderBookChart({
         },
         resolveSymbol: (
           symbolName,
-          onSymbolResolvedCallback,
-          onResolveErrorCallback,
-          extension
+          onSymbolResolvedCallback
+          // onResolveErrorCallback,
+          // extension
         ) => {
           // if not on correct page, go to correct page
           navigate(`/orderbook/${symbolName}`);
@@ -208,8 +208,8 @@ export default function OrderBookChart({
         },
         searchSymbols: (
           userInput = '',
-          exchange, // will always be "Duality"
-          symbolType, // will always be "crypto"
+          _exchange, // will always be "Duality"
+          _symbolType, // will always be "crypto"
           onResultReadyCallback
         ) => {
           const tokens = Array.from(new Set(tokenPairs.flatMap((v) => v)));
@@ -288,8 +288,8 @@ export default function OrderBookChart({
           symbolInfo,
           resolution,
           onRealtimeCallback,
-          subscriberUID,
-          onResetCacheNeededCallback
+          subscriberUID
+          // onResetCacheNeededCallback
         ) => {
           const fetchID = getFetchID(symbolInfo, resolution);
           const url = getFetchURL(tokenIdA, tokenIdB, resolution, {
