+++ src/pages/Pool/MyPositionTableCard.tsx
@@ -18,8 +18,6 @@ import TableCard from '../../components/cards/TableCard';
 import ValueBar from '../../components/Table/ValueBar';
 
 function MyPositionTableCard({
-  tokenA,
-  tokenB,
   title = 'My Position',
   header,
   data,
@@ -248,7 +246,7 @@ export function MyEditedPositionTableCard({
       editedUserPosition
         // sort by price
         .sort((a, b) => {
-          return !!invertedTokenOrder
+          return invertedTokenOrder
             ? b.deposit.centerTickIndex.toNumber() -
                 a.deposit.centerTickIndex.toNumber()
             : a.deposit.centerTickIndex.toNumber() -
